requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
idna==2.6
numpy==1.13.3
pandas==0.20.3
pandas-datareader==0.5.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
six==1.11.0
urllib3==1.22
xlwt==1.3.0
